backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
requests
language-tool-python
vaderSentiment
python-multipart